images: move image package to images
